+++ packages/at_client/lib/src/client/at_client_impl.dart
@@ -3,25 +3,19 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:at_base2e15/at_base2e15.dart';
-import 'package:at_client/src/client/at_client_spec.dart';
-import 'package:at_client/src/client/local_secondary.dart';
-import 'package:at_client/src/client/remote_secondary.dart';
+import 'package:at_client/at_client.dart';
 import 'package:at_client/src/client/secondary.dart';
 import 'package:at_client/src/client/verb_builder_manager.dart';
 import 'package:at_client/src/compaction/at_commit_log_compaction.dart';
-import 'package:at_client/src/encryption_service/encryption_manager.dart';
 import 'package:at_client/src/listener/at_sign_change_listener.dart';
 import 'package:at_client/src/listener/switch_at_sign_event.dart';
-import 'package:at_client/src/manager/at_client_manager.dart';
 import 'package:at_client/src/manager/storage_manager.dart';
 import 'package:at_client/src/manager/sync_manager.dart';
 import 'package:at_client/src/manager/sync_manager_impl.dart';
 import 'package:at_client/src/preference/at_client_config.dart';
-import 'package:at_client/src/preference/at_client_preference.dart';
 import 'package:at_client/src/response/response.dart';
 import 'package:at_client/src/service/encryption_service.dart';
 import 'package:at_client/src/service/file_transfer_service.dart';
-import 'package:at_client/src/service/notification_service.dart';
 import 'package:at_client/src/service/sync_service.dart';
 import 'package:at_client/src/stream/at_stream_notification.dart';
 import 'package:at_client/src/stream/at_stream_response.dart';
@@ -31,13 +25,10 @@ import 'package:at_client/src/transformer/request_transformer/get_request_transf
 import 'package:at_client/src/transformer/request_transformer/put_request_transformer.dart';
 import 'package:at_client/src/transformer/response_transformer/get_response_transformer.dart';
 import 'package:at_client/src/transformer/response_transformer/put_response_transformer.dart';
-import 'package:at_client/src/util/at_client_util.dart';
 import 'package:at_client/src/util/at_client_validation.dart';
 import 'package:at_client/src/util/constants.dart';
 import 'package:at_client/src/util/sync_util.dart';
-import 'package:at_client/src/client/request_options.dart';
 import 'package:at_commons/at_builders.dart';
-import 'package:at_commons/at_commons.dart';
 import 'package:at_lookup/at_lookup.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
 import 'package:at_utils/at_utils.dart';
@@ -45,6 +36,7 @@ import 'package:at_chops/at_chops.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart';
 import 'package:uuid/uuid.dart';
+import 'package:version/version.dart';
 
 /// Implementation of [AtClient] interface and [AtSignChangeListener] interface
 ///
@@ -136,7 +128,7 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
       AtClientCommitLogCompaction? atClientCommitLogCompaction,
       AtClientConfig? atClientConfig}) async {
     atClientManager ??= AtClientManager.getInstance();
-    currentAtSign = AtUtils.formatAtSign(currentAtSign)!;
+    currentAtSign = AtUtils.fixAtSign(currentAtSign);
 
     // Fetch cached AtClientImpl for re-use, or create a new one and init it
     AtClientImpl? atClientImpl;
@@ -170,7 +162,7 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
       AtChops? atChops,
       AtClientCommitLogCompaction? atClientCommitLogCompaction,
       AtClientConfig? atClientConfig}) {
-    _atSign = AtUtils.formatAtSign(theAtSign)!;
+    _atSign = AtUtils.fixAtSign(theAtSign);
     _logger = AtSignLogger('AtClientImpl ($_atSign)');
     _preference = preference;
     _preference?.namespace ??= namespace;
@@ -242,17 +234,6 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
       return _localSecondary!;
     }
     return _remoteSecondary!;
-  }
-
-  @override
-  Future<void> startMonitor(String privateKey, Function? notificationCallback,
-      {String? regex}) async {
-    var monitorVerbBuilder = MonitorVerbBuilder();
-    if (regex != null) {
-      monitorVerbBuilder.regex = regex;
-    }
-    await _remoteSecondary!.monitor(
-        monitorVerbBuilder.buildCommand(), notificationCallback, privateKey);
   }
 
   @override
@@ -417,7 +398,7 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
 
   @override
   Future<bool> put(AtKey atKey, dynamic value,
-      {bool isDedicated = false}) async {
+      {bool isDedicated = false, PutRequestOptions? putRequestOptions}) async {
     _telemetry?.controller.sink
         .add(AtTelemetryEvent('AtClient.put called', {"key": atKey}));
     // If the value is neither String nor List<int> throw exception
@@ -427,10 +408,12 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
     }
     AtResponse atResponse = AtResponse();
     if (value is String) {
-      atResponse = await putText(atKey, value);
+      atResponse =
+          await putText(atKey, value, putRequestOptions: putRequestOptions);
     }
     if (value is List<int>) {
-      atResponse = await putBinary(atKey, value);
+      atResponse =
+          await putBinary(atKey, value, putRequestOptions: putRequestOptions);
     }
     _telemetry?.controller.sink
         .add(AtTelemetryEvent('AtClient.put complete', {"atKey": atKey}));
@@ -439,13 +422,14 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
 
   /// put's the text data into the keystore
   @override
-  Future<AtResponse> putText(AtKey atKey, String value) async {
+  Future<AtResponse> putText(AtKey atKey, String value,
+      {PutRequestOptions? putRequestOptions}) async {
     try {
       // Set the default metadata if not already set.
       atKey.metadata ??= Metadata();
       // Setting metadata.isBinary to false for putText
       atKey.metadata!.isBinary = false;
-      return await _putInternal(atKey, value);
+      return await _putInternal(atKey, value, putRequestOptions);
     } on AtException catch (e) {
       throw AtExceptionManager.createException(e);
     }
@@ -453,14 +437,16 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
 
   /// put's the binary data(e.g. images, files etc) into the keystore
   @override
-  Future<AtResponse> putBinary(AtKey atKey, List<int> value) async {
+  Future<AtResponse> putBinary(AtKey atKey, List<int> value,
+      {PutRequestOptions? putRequestOptions}) async {
     try {
       // Set the default metadata if not already set.
       atKey.metadata ??= Metadata();
       // Setting metadata.isBinary to true for putBinary
       atKey.metadata!.isBinary = true;
       // Base2e15.encode method converts the List<int> type to String.
-      return await _putInternal(atKey, Base2e15.encode(value));
+      return await _putInternal(
+          atKey, Base2e15.encode(value), putRequestOptions);
     } on AtException catch (e) {
       throw AtExceptionManager.createException(e);
     }
@@ -477,7 +463,8 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
     }
   }
 
-  Future<AtResponse> _putInternal(AtKey atKey, dynamic value) async {
+  Future<AtResponse> _putInternal(
+      AtKey atKey, dynamic value, PutRequestOptions? putRequestOptions) async {
     // Performs the put request validations.
     AtClientValidation.validatePutRequest(atKey, value, preference!);
     // Set sharedBy to currentAtSign if not set.
@@ -488,6 +475,9 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
       atKey.namespace ??= preference?.namespace;
     }
 
+    if (preference!.atProtocolEmitted >= Version(2, 0, 0)) {
+      atKey.metadata!.ivNonce ??= EncryptionUtil.generateIV();
+    }
     ensureLowerCase(atKey);
 
     // validate the atKey
@@ -521,7 +511,8 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
     // Transform put request
     // Optionally passing encryption private key to sign the public data.
     UpdateVerbBuilder verbBuilder = await putRequestTransformer.transform(tuple,
-        encryptionPrivateKey: encryptionPrivateKey);
+        encryptionPrivateKey: encryptionPrivateKey,
+        requestOptions: putRequestOptions);
     // Validate the size of the value after encryption/encoding
     // Since AtClientPreference is mandatory argument in create method, _preference
     // will not be null.
@@ -537,44 +528,6 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
       return AtResponse()..isError = true;
     }
     return await PutResponseTransformer().transform(putResponse);
-  }
-
-  @override
-  Future<bool> notify(AtKey atKey, String value, OperationEnum operation,
-      {MessageTypeEnum? messageType,
-      PriorityEnum? priority,
-      StrategyEnum? strategy,
-      int? latestN,
-      String? notifier = SYSTEM,
-      bool isDedicated = false}) async {
-    AtKeyValidators.get().validate(
-        atKey.toString(),
-        ValidationContext()
-          ..atSign = _atSign
-          ..validateOwnership = true);
-    final notificationParams =
-        NotificationParams.forUpdate(atKey, value: value);
-    final notifyResult = await notificationService.notify(notificationParams);
-    return notifyResult.notificationStatusEnum ==
-        NotificationStatusEnum.delivered;
-  }
-
-  @override
-  Future<String> notifyAll(AtKey atKey, String value, OperationEnum operation,
-      {bool isDedicated = false}) async {
-    var returnMap = {};
-    var sharedWithList = jsonDecode(atKey.sharedWith!);
-    for (var sharedWith in sharedWithList) {
-      atKey.sharedWith = sharedWith;
-      final notificationParams =
-          NotificationParams.forUpdate(atKey, value: value);
-      final notifyResult = await notificationService.notify(notificationParams);
-      returnMap.putIfAbsent(
-          sharedWith,
-          () => (notifyResult.notificationStatusEnum ==
-              NotificationStatusEnum.delivered));
-    }
-    return jsonEncode(returnMap);
   }
 
   @override
@@ -877,7 +830,8 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
         var decryptedFile = await _encryptionService!.decryptFileInChunks(
             File(encryptedFile.path),
             fileTransferObject.fileEncryptionKey,
-            _preference!.fileEncryptionChunkSize);
+            _preference!.fileEncryptionChunkSize,
+            ivBase64: fileTransferObject.ivBase64);
         decryptedFile.copySync(downloadPath +
             Platform.pathSeparator +
             encryptedFile.path.split(Platform.pathSeparator).last);
@@ -904,107 +858,6 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
   }
 
   @override
-  Future<String?> notifyChange(NotificationParams notificationParams) async {
-    String? notifyKey = notificationParams.atKey.key;
-
-    // validate sharedWith atSign
-    AtUtils.fixAtSign(notificationParams.atKey.sharedWith!);
-    // Check if sharedWith AtSign exists
-    await AtClientValidation().isAtSignExists(
-        _atClientManager.secondaryAddressFinder!,
-        notificationParams.atKey.sharedWith!);
-    // validate sharedBy atSign
-    if (notificationParams.atKey.sharedBy == null ||
-        notificationParams.atKey.sharedBy!.isEmpty) {
-      notificationParams.atKey.sharedBy = getCurrentAtSign();
-    }
-    AtUtils.fixAtSign(notificationParams.atKey.sharedBy!);
-    // validate atKey
-    // For messageType is text, text may contains spaces but key should not have spaces
-    // Hence do not validate the key.
-    if (notificationParams.messageType != MessageTypeEnum.text) {
-      AtClientValidation.validateKey(notificationParams.atKey.key);
-      ValidationResult validationResult = AtKeyValidators.get().validate(
-          notificationParams.atKey.toString(),
-          ValidationContext()
-            ..atSign = _atSign
-            ..validateOwnership = true);
-      if (!validationResult.isValid) {
-        throw AtClientException('AT0014', validationResult.failureReason);
-      }
-    }
-    // validate metadata
-    AtClientValidation.validateMetadata(notificationParams.atKey.metadata);
-    // If namespaceAware is set to true, append nameSpace to key.
-    if (notificationParams.atKey.metadata != null &&
-        notificationParams.atKey.metadata!.namespaceAware) {
-      notifyKey = _getKeyWithNamespace(notifyKey!);
-    }
-    notificationParams.atKey.sharedBy ??= _atSign;
-
-    var builder = NotifyVerbBuilder()
-      ..id = notificationParams.id
-      ..atKey = notifyKey
-      ..sharedBy = notificationParams.atKey.sharedBy
-      ..sharedWith = notificationParams.atKey.sharedWith
-      ..operation = notificationParams.operation
-      ..messageType = notificationParams.messageType
-      ..priority = notificationParams.priority
-      ..strategy = notificationParams.strategy
-      ..latestN = notificationParams.latestN
-      ..notifier = notificationParams.notifier;
-    // if no metadata is set by the app, init empty metadata
-    notificationParams.atKey.metadata ??= Metadata();
-    // If value is not null, encrypt the value
-    if (notificationParams.value != null &&
-        notificationParams.value!.isNotEmpty) {
-      // If atKey is being notified to another atSign, encrypt data with other
-      // atSign encryption public key.
-      if (notificationParams.atKey.sharedWith != null &&
-          notificationParams.atKey.sharedWith != _atSign) {
-        try {
-          final atKeyEncryption = AtKeyEncryptionManager(this)
-              .get(notificationParams.atKey, _atSign);
-          builder.value = await atKeyEncryption.encrypt(
-              notificationParams.atKey, notificationParams.value!);
-        } on KeyNotFoundException catch (e) {
-          return Future.error(
-              AtClientException(error_codes['AtClientException'], e.message));
-        }
-      }
-      // If sharedWith is currentAtSign, encrypt data with currentAtSign encryption public key.
-      if (notificationParams.atKey.sharedWith == null ||
-          notificationParams.atKey.sharedWith == _atSign) {
-        try {
-          final atKeyEncryption = AtKeyEncryptionManager(this)
-              .get(notificationParams.atKey, _atSign);
-          builder.value = await atKeyEncryption.encrypt(
-              notificationParams.atKey, notificationParams.value!);
-        } on KeyNotFoundException catch (e) {
-          return Future.error(
-              AtClientException(error_codes['AtClientException'], e.message));
-        }
-      }
-    }
-    // If metadata is not null, add metadata to notify builder object.
-    if (notificationParams.atKey.metadata != null) {
-      builder.ttl = notificationParams.atKey.metadata!.ttl;
-      builder.ttb = notificationParams.atKey.metadata!.ttb;
-      builder.ttr = notificationParams.atKey.metadata!.ttr;
-      builder.ccd = notificationParams.atKey.metadata!.ccd;
-      builder.isPublic = notificationParams.atKey.metadata!.isPublic!;
-      builder.sharedKeyEncrypted =
-          notificationParams.atKey.metadata!.sharedKeyEnc;
-      builder.pubKeyChecksum = notificationParams.atKey.metadata!.pubKeyCS;
-    }
-    if (notifyKey!.startsWith(AT_PKAM_PRIVATE_KEY) ||
-        notifyKey.startsWith(AT_PKAM_PUBLIC_KEY)) {
-      builder.sharedBy = null;
-    }
-    return await getRemoteSecondary()?.executeVerb(builder);
-  }
-
-  @override
   void listenToAtSignChange(SwitchAtSignEvent switchAtSignEvent) {
     // Checks if the instance of AtClientImpl belongs to previous atSign. If Yes,
     // the compaction job is stopped and removed from changeListener list.
@@ -1015,10 +868,78 @@ class AtClientImpl implements AtClient, AtSignChangeListener {
     }
   }
 
-  // TODO v4 - remove this method in version 4 of at_client package
+  // TODO v4 - remove the follow methods in version 4 of at_client package
+
   @override
   @Deprecated("Use AtClient.syncService")
   SyncManager? getSyncManager() {
     return SyncManagerImpl.getInstance().getSyncManager(_atSign);
+  }
+
+  @override
+
+  ///[Deprecated] Use [AtClient.notificationService]
+  @Deprecated('Use AtClient.notificationService')
+  Future<void> startMonitor(String privateKey, Function? notificationCallback,
+      {String? regex}) async {
+    var monitorVerbBuilder = MonitorVerbBuilder();
+    if (regex != null) {
+      monitorVerbBuilder.regex = regex;
+    }
+    await _remoteSecondary!.monitor(
+        monitorVerbBuilder.buildCommand(), notificationCallback, privateKey);
+  }
+
+  @override
+  @Deprecated("Use NotificationService")
+  Future<bool> notify(AtKey atKey, String value, OperationEnum operation,
+      {MessageTypeEnum? messageType,
+      PriorityEnum? priority,
+      StrategyEnum? strategy,
+      int? latestN,
+      String? notifier = SYSTEM,
+      bool isDedicated = false}) async {
+    AtKeyValidators.get().validate(
+        atKey.toString(),
+        ValidationContext()
+          ..atSign = _atSign
+          ..validateOwnership = true);
+    final notificationParams =
+        NotificationParams.forUpdate(atKey, value: value);
+    final notifyResult = await notificationService.notify(notificationParams);
+    return notifyResult.notificationStatusEnum ==
+        NotificationStatusEnum.delivered;
+  }
+
+  @override
+  @Deprecated('Use NotificationService')
+  Future<String> notifyAll(AtKey atKey, String value, OperationEnum operation,
+      {bool isDedicated = false}) async {
+    var returnMap = {};
+    var sharedWithList = jsonDecode(atKey.sharedWith!);
+    for (var sharedWith in sharedWithList) {
+      atKey.sharedWith = sharedWith;
+      final notificationParams =
+          NotificationParams.forUpdate(atKey, value: value);
+      final notifyResult = await notificationService.notify(notificationParams);
+      returnMap.putIfAbsent(
+          sharedWith,
+          () => (notifyResult.notificationStatusEnum ==
+              NotificationStatusEnum.delivered));
+    }
+    return jsonEncode(returnMap);
+  }
+
+  @override
+
+  ///[Deprecated] Use [NotificationService.notify]
+  @Deprecated("Use [NotificationService.notify]")
+  Future<String?> notifyChange(NotificationParams notificationParams) async {
+    NotificationResult result =
+        await notificationService.notify(notificationParams);
+    if (result.atClientException != null) {
+      throw result.atClientException!;
+    }
+    return result.notificationID;
   }
 }