+++ packages/at_client/lib/src/response/at_notification.dart
@@ -11,18 +11,36 @@ class AtNotification {
   String? operation;
   String? messageType;
   bool? isEncrypted;
+  int? expiresAtInEpochMillis;
+  Metadata? metadata;
 
   /// AtNotification instance is created without initializing the fields
   AtNotification.empty();
 
   AtNotification(this.id, this.key, this.from, this.to, this.epochMillis,
       this.messageType, this.isEncrypted,
-      {this.value, this.operation});
+      {this.value, this.operation, this.expiresAtInEpochMillis, this.metadata});
 
   factory AtNotification.fromJson(Map<String, dynamic> json) {
+    Metadata? metadata;
+
+    if (json['metadata'] != null) {
+      metadata = Metadata();
+      metadata.encKeyName = json['metadata'][ENCRYPTING_KEY_NAME];
+      metadata.encAlgo = json['metadata'][ENCRYPTING_ALGO];
+      metadata.ivNonce = json['metadata'][IV_OR_NONCE];
+      metadata.skeEncKeyName =
+          json['metadata'][SHARED_KEY_ENCRYPTED_ENCRYPTING_KEY_NAME];
+      metadata.skeEncAlgo =
+          json['metadata'][SHARED_KEY_ENCRYPTED_ENCRYPTING_ALGO];
+    }
+
     return AtNotification(json['id'], json['key'], json['from'], json['to'],
         json['epochMillis'], json['messageType'], json[IS_ENCRYPTED],
-        value: json['value'], operation: json['operation']);
+        value: json['value'],
+        operation: json['operation'],
+        expiresAtInEpochMillis: json['expiresAt'],
+        metadata: metadata);
   }
 
   Map<String, dynamic> toJson() {
@@ -36,7 +54,9 @@ class AtNotification {
       'operation': operation,
       'messageType': messageType,
       IS_ENCRYPTED: isEncrypted,
-      'notificationStatus': status
+      'notificationStatus': status,
+      'expiresAt': expiresAtInEpochMillis,
+      'metadata': metadata
     };
   }
 
@@ -51,6 +71,6 @@ class AtNotification {
 
   @override
   String toString() {
-    return 'AtNotification{id: $id, key: $key, from: $from, to: $to, epochMillis: $epochMillis, value: $value, operation: $operation, status: $status}';
+    return toJson().toString();
   }
 }